fix: bug with import styles
