+++ packages/client/src/pages/auth/styled.ts
@@ -1,4 +1,5 @@
-import { Button, Form, Typography } from 'antd'
+import { Button, Form, Space, Typography } from 'antd'
+import { Link } from 'react-router-dom'
 import styled from 'styled-components'
 
 export const AuthForm = styled(Form)`
@@ -20,4 +21,23 @@ export const AuthTitle = styled(Typography.Title)`
 
 export const AuthSubmitButton = styled(Button)`
   margin-top: 15px;
+  width: 100%;
+`
+
+export const AuthFooterText = styled(Typography.Text)`
+  font-size: 14px;
+  color: rgba(0, 0, 0, 0.45);
+`
+
+export const AuthLink = styled(Link)`
+  color: #1890ff;
+  &:hover {
+    color: #40a9ff;
+  }
+`
+
+export const AuthSpace = styled(Space)`
+  width: 100%;
+  justify-content: center;
+  margin-top: 16px;
 `
